Add a mappable action to cleanup shared SSH connections
